The amount of RAM buffer pool is configurable
Some system doesn't have much memory.
For the same reason segment_size_order became configurable
this should be parameterized.
The default 64MB is surpressing in a 32bit system.

Also, segment_size_order and this value
are optional now.
Added ARG_EXIST macro for this purpose.

If the RAM buffer pool became resizable
is the question remains.
Do we remove this parameter in that case?
Maybe setting the amount to 2MB is enough.