BUGFIX: cursor rotation is delayed
BUG_ON() in prepare_segment_header_device() now doesn't incur.
rotating the cursor immediately makes an exception at the very end of
the log. To remove the exception, the concept of cursor is extended so
that it can place in one-step outside the boundary (like sentinel).

This way, removes the exception.

Signed-off-by: Akira Hayakawa <ruby.wktk@gmail.com>