severely buggy!!! Let's solve it!
mkfs.ext4 doesn't pass with plog but does without plog.
even with plog it does if I use queue_barrier_io() to defer
instead of simply ack by bio_endio() or simply deferring the ack.
It seems to be the difference is that the data on RAM buffer is
flushed to the cache and RAM buffer gets cleaned up after REQ_FLUSH is
acked in case of using queue_barrier_io(). This makes read hits right
after REQ_FLUSH is hit on the cache not on the RAM buffer. On the other
hand, in the failure case, it hits on the RAM buffer.

I see read hits on difference locations (RAM buffer or cache device) in
these two case and assume RAM buffer hit case is buggy.

Subsequent commits will solve this problem.

Signed-off-by: Akira Hayakawa <ruby.wktk@gmail.com>