FIX: async I/O should not be wrapped by IO() macro
this macro disable I/O if the device dead
but what if the I/O dec some refcount taken on the callback?
and what if the callback wakes up the foreground when the refcount
falls to zero?

This patch removes all IO macro wrappers around async I/O in migration
and the plog write.

As the result, regression test for type 1 runs through.

Signed-off-by: Akira Hayakawa <ruby.wktk@gmail.com>